[tests] linting
